+++ src/main/resources/META-INF/services/org.keycloak.broker.social.SocialIdentityProviderFactory
@@ -1,2 +1,2 @@
-fr.insee.keycloak.provider.FranceConnectIdentityProviderFactory
-fr.insee.keycloak.provider.AgentConnectIdentityProviderFactory
+fr.insee.keycloak.providers.franceconnect.FranceConnectIdentityProviderFactory
+fr.insee.keycloak.providers.agentconnect.AgentConnectIdentityProviderFactory
